requirements.txt
langchain_chroma
langchain_core
langchain_openai
openai
langchain
python-dotenv
gradio

